+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 Change-Log
 ===============
 
+### Release 1.0.1 ###
+> --.02.2017
+
+- `BaseNavigationalTransition` now allows to specify `Bundle` with extras for the transition activity
+  via `BaseNavigationalTransition.intentExtras(Bundle)` method.
+
 ### Release 1.0.0 ###
 > 19.01.2017